Imported zero points with the correct integer format

+++ python/tvm/relax/frontend/tflite/tflite_frontend.py
@@ -985,7 +985,7 @@ class OperatorConverter:
                 if is_qnn_params_valid:
                     qnn_params = dict()
                     qnn_params["scale"] = relax.const(scale, "float32")
-                    qnn_params["zero_point"] = relax.const(zero_point, "int32")
+                    qnn_params["zero_point"] = relax.const(zero_point, "int8")
                     qnn_params["axis"] = int(tflite_qnn_params.QuantizedDimension())
             return_list.append(TensorWrapper(tensor_idx, tensor, buffer, qnn_params))
         return return_list
